Move useRecommendedChannels to shared hooks directory.
