+++ plugins/woocommerce-admin/client/marketing/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useInstalledPlugins } from './useInstalledPlugins';
+export { useRecommendedChannels } from './useRecommendedChannels';
